Move EhCacheProviderWithXMLOrJavaBasedConfig(Test) to org.ehcache.impl.config.persistence (#2876)
