+++ impl/src/test/java/org/ehcache/impl/config/persistence/EhCacheProviderWithXMLOrJavaBasedConfigTest.java
@@ -1,7 +1,7 @@
-package org.ehcache.core.config;
+package org.ehcache.impl.config.persistence;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
 
 import java.io.IOException;
 import java.nio.file.Files;
