+++ impl/src/test/java/org/ehcache/impl/config/persistence/ehcache-dir-persistence-test.xml
